chore: Move Zapier settings page to plugin
