+++ plugins/zapier/client/Settings.tsx
@@ -1,20 +1,22 @@
 import * as React from "react";
 import { Helmet } from "react-helmet-async";
 import { Trans } from "react-i18next";
+import { IntegrationScene } from "~/scenes/Settings/components/IntegrationScene";
 import Heading from "~/components/Heading";
-import ZapierIcon from "~/components/Icons/ZapierIcon";
-import Scene from "~/components/Scene";
 import Text from "~/components/Text";
 import env from "~/env";
+import useCurrentUser from "~/hooks/useCurrentUser";
 import useStores from "~/hooks/useStores";
+import ZapierIcon from "./Icon";
 
 function Zapier() {
   const { ui } = useStores();
+  const user = useCurrentUser();
   const { resolvedTheme } = ui;
   const appName = env.APP_NAME;
 
   return (
-    <Scene title="Zapier" icon={<ZapierIcon />}>
+    <IntegrationScene title="Zapier" icon={<ZapierIcon />}>
       <Heading>Zapier</Heading>
       <Helmet>
         <script
@@ -39,14 +41,15 @@ function Zapier() {
       <zapier-app-directory
         app="outline"
         link-target="new-tab"
-        theme={resolvedTheme}
+        sign-up-email={user.email}
+        theme={resolvedTheme === "system" ? undefined : resolvedTheme}
         hide="notion,confluence-cloud,confluence,google-docs,slack"
         applimit={6}
         introcopy="hide"
         create-without-template="show"
         use-this-zap="show"
       />
-    </Scene>
+    </IntegrationScene>
   );
 }
