+++ shared/i18n/locales/en_US/translation.json
@@ -528,8 +528,8 @@
   "Applications": "Applications",
   "Shared Links": "Shared Links",
   "Import": "Import",
-  "Integrations": "Integrations",
   "Install": "Install",
+  "Integrations": "Integrations",
   "Revoke token": "Revoke token",
   "Revoke": "Revoke",
   "Show path to document": "Show path to document",
@@ -1122,7 +1122,6 @@
   "You can create templates to help your team create consistent and accurate documentation.": "You can create templates to help your team create consistent and accurate documentation.",
   "Alphabetical": "Alphabetical",
   "There are no templates just yet.": "There are no templates just yet.",
-  "Zapier is a platform that allows {{appName}} to easily integrate with thousands of other business tools. Automate your workflows, sync data, and more.": "Zapier is a platform that allows {{appName}} to easily integrate with thousands of other business tools. Automate your workflows, sync data, and more.",
   "A confirmation code has been sent to your email address, please enter the code below to permanently destroy this workspace.": "A confirmation code has been sent to your email address, please enter the code below to permanently destroy this workspace.",
   "Confirmation code": "Confirmation code",
   "Deleting the <1>{{workspaceName}}</1> workspace will destroy all collections, documents, users, and associated data. You will be immediately logged out of {{appName}}.": "Deleting the <1>{{workspaceName}}</1> workspace will destroy all collections, documents, users, and associated data. You will be immediately logged out of {{appName}}.",
@@ -1218,6 +1217,7 @@
   "Webhooks can be used to notify your application when events happen in {{appName}}. Events are sent as a https request with a JSON payload in near real-time.": "Webhooks can be used to notify your application when events happen in {{appName}}. Events are sent as a https request with a JSON payload in near real-time.",
   "Inactive": "Inactive",
   "Create a webhook": "Create a webhook",
+  "Zapier is a platform that allows {{appName}} to easily integrate with thousands of other business tools. Automate your workflows, sync data, and more.": "Zapier is a platform that allows {{appName}} to easily integrate with thousands of other business tools. Automate your workflows, sync data, and more.",
   "Never logged in": "Never logged in",
   "Online now": "Online now",
   "Online {{ timeAgo }}": "Online {{ timeAgo }}",
